+++ EasySequence/Classes/EasySequence.h
@@ -32,6 +32,10 @@
 #import <EasySequence/NSSet+EZSTransfer.h>
 #import <EasySequence/NSOrderedSet+EZSTransfer.h>
 
+// Utils
+#import <EasySequence/EZSWeakReference.h>
+#import <EasySequence/NSObject+EZSDeallocBell.h>
+
 FOUNDATION_EXPORT double EasySequenceVersionNumber;
 FOUNDATION_EXPORT const unsigned char EasySequenceVersionString[];