+++ EasySequence/Classes/Sequences/EZSWeakOrderedSet.m
@@ -23,7 +23,14 @@
 }
 
 - (NSArray *)allObjects {
-    return [[EZS_Sequence([super allObjects]) map:EZS_propertyWith(EZS_KeyPath(EZSWeakReference, reference))] as:NSArray.class];
+    NSMutableArray *array = [NSMutableArray array];
+    for (EZSWeakReference *reference in [super allObjects]) {
+        id strongItem = reference.reference;
+        if (strongItem) {
+            [array addObject:strongItem];
+        }
+    }
+    return array;
 }
 
 - (EZSWeakReference *)weakReference:(id _Nonnull)anObject {