+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 
 # Change Log
 
+## 1.1.0 
+
+- Add Generic support for EZSWeakArray, EZSWeakOrderedSet.
+- Add EZSWeakReference and NSObject+EZSDeallocBell helper classes.
+
 ## 1.0.0
 
 - First version. See also README.md.